// packages/schemas/src/nfe/PL_010e.ts
export type TNFe = {
  /** Informações da Nota Fiscal eletrônica */
  infNFe: TNFe_infNFe;
  /** Informações suplementares Nota Fiscal */
  infNFeSupl?: TNFe_infNFeSupl;
  Signature: SignatureType;
};
// ...
export type TProtNFe = {
  /** @attribute xsd:TVerNFe, pattern `4\.00` */
  versao: string;
  /** Dados do protocolo de status */
  infProt: TProtNFe_infProt;
  Signature?: SignatureType;
};

/**
 * Tipo da NF-e processada
 * xsd: TNfeProc
 */
export type TNfeProc = {
  /** @attribute xsd:TVerNFe, pattern `4\.00` */
  versao: string;
  NFe: TNFe;
  protNFe: TProtNFe;
};
// ...
export const TNFe: ComplexType<TNFe> = { id: "TNFe", ns: "http://www.portalfiscal.inf.br/nfe", c: { g: "s", i: [{ e: "infNFe", t: TNFe_infNFe, u: ["nItem"] }, { e: "infNFeSupl", t: TNFe_infNFeSupl, n: 0 }, { e: "Signature", t: SignatureType, ns: "http://www.w3.org/2000/09/xmldsig#" }] } };
// ...
export const TProtNFe: ComplexType<TProtNFe> = { id: "TProtNFe", ns: "http://www.portalfiscal.inf.br/nfe", a: [{ a: "versao", t: st_TVerNFe, r: 1 }], c: { g: "s", i: [{ e: "infProt", t: TProtNFe_infProt }, { e: "Signature", t: SignatureType, n: 0, ns: "http://www.w3.org/2000/09/xmldsig#" }] } };
export const TNfeProc: ComplexType<TNfeProc> = { id: "TNfeProc", ns: "http://www.portalfiscal.inf.br/nfe", a: [{ a: "versao", t: st_TVerNFe, r: 1 }], c: { g: "s", i: [{ e: "NFe", t: TNFe }, { e: "protNFe", t: TProtNFe }] } };
// ...
export const nfeProcElement: ElementoRaiz<TNfeProc> = { nome: "nfeProc", ns: "http://www.portalfiscal.inf.br/nfe", tipo: TNfeProc };